browser-use/pyproject.toml
[project]
name = "browser-use-scraper"
version = "0.1.0"
description = "A project to explore Browser use and it's capabilitites"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use[memory]>=0.1.45",
    "langchain-openai>=0.3.11",
    "python-dotenv>=1.1.0",
]
